--- a/aiogram/methods/base.py
+++ b/aiogram/methods/base.py
@@ -1,13 +1,22 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Any, Dict, Generator, Generic, Optional, TypeVar
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Dict,
+    Generator,
+    Generic,
+    Optional,
+    TypeVar,
+    Union,
+)
 
-from pydantic import BaseConfig, BaseModel, Extra, root_validator
-from pydantic.generics import GenericModel
+import msgspec
+
+from aiogram.types import *  # noqa
 
 from ..types import InputFile, ResponseParameters
-from ..types.base import UNSET_TYPE
 
 if TYPE_CHECKING:
     from ..client.bot import Bot
@@ -15,44 +24,30 @@ if TYPE_CHECKING:
 TelegramType = TypeVar("TelegramType", bound=Any)
 
 
-class Request(BaseModel):
+class Request(msgspec.Struct, weakref=True):
     method: str
 
     data: Dict[str, Optional[Any]]
     files: Optional[Dict[str, InputFile]]
 
-    class Config(BaseConfig):
-        arbitrary_types_allowed = True
 
-
-class Response(GenericModel, Generic[TelegramType]):
+class Response(msgspec.Struct, Generic[TelegramType], weakref=True, kw_only=True):
     ok: bool
-    result: Optional[TelegramType] = None
+    result: Optional[Union[Any, None]] = None
     description: Optional[str] = None
     error_code: Optional[int] = None
     parameters: Optional[ResponseParameters] = None
 
 
-class TelegramMethod(ABC, BaseModel, Generic[TelegramType]):
-    class Config(BaseConfig):
-        # use_enum_values = True
-        extra = Extra.allow
-        allow_population_by_field_name = True
-        arbitrary_types_allowed = True
-        orm_mode = True
-        smart_union = True  # https://github.com/aiogram/aiogram/issues/901
-
-    @root_validator(pre=True)
-    def remove_unset(cls, values: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Remove UNSET before fields validation.
-
-        We use UNSET as a sentinel value for `parse_mode` and replace it to real value later.
-        It isn't a problem when it's just default value for a model field,
-        but UNSET might be passing to a model initialization from `Bot.method_name`,
-        so we must take care of it and remove it before fields validation.
-        """
-        return {k: v for k, v in values.items() if not isinstance(v, UNSET_TYPE)}
+class TelegramMethod(
+    msgspec.Struct,
+    Generic[TelegramType],
+    omit_defaults=True,
+    weakref=True,
+    kw_only=True,
+    forbid_unknown_fields=False,
+):
+    method: str = msgspec.UNSET
 
     @property
     @abstractmethod
@@ -64,15 +59,13 @@ class TelegramMethod(ABC, BaseModel, Generic[TelegramType]):
     def __api_method__(self) -> str:
         pass
 
-    def dict(self, **kwargs: Any) -> Any:
-        # override dict of pydantic.BaseModel to overcome exporting request_timeout field
-        exclude = kwargs.pop("exclude", set())
-
-        return super().dict(exclude=exclude, **kwargs)
-
-    def build_response(self, data: Dict[str, Any]) -> Response[TelegramType]:
+    def build_response(self, data: Union[str, bytes, Dict[str, Any]]) -> Response[TelegramType]:
         # noinspection PyTypeChecker
-        return Response[self.__returning__](**data)  # type: ignore
+        if isinstance(data, (bytes, str)):
+            data = msgspec.json.decode(data)
+        if "result" in data:
+            data["result"] = msgspec.from_builtins(data["result"], type=self.__returning__)
+        return msgspec.from_builtins(data, type=Response)
 
     async def emit(self, bot: Bot) -> TelegramType:
         return await bot(self)