--- a/aiogram/methods/forward_message.py
+++ b/aiogram/methods/forward_message.py
@@ -2,6 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Optional, Union
 
+import msgspec
+
 from ..types import Message
 from ..types.base import UNSET_PROTECT_CONTENT
 from .base import TelegramMethod
@@ -27,5 +29,5 @@ class ForwardMessage(TelegramMethod[Message]):
     """Unique identifier for the target message thread (topic) of the forum; for forum supergroups only"""
     disable_notification: Optional[bool] = None
     """Sends the message `silently <https://telegram.org/blog/channels-2-0#silent-messages>`_. Users will receive a notification with no sound."""
-    protect_content: Optional[bool] = UNSET_PROTECT_CONTENT
+    protect_content: Optional[bool] = msgspec.field(default_factory=lambda: UNSET_PROTECT_CONTENT)
     """Protects the contents of the forwarded message from forwarding and saving"""