… aiogram/methods/send_game.py
… aiogram/methods/send_invoice.py
… aiogram/methods/send_location.py
… aiogram/methods/send_media_group.py
… aiogram/methods/send_message.py
… aiogram/methods/send_photo.py
… aiogram/methods/send_poll.py
… aiogram/methods/send_sticker.py
… aiogram/methods/send_venue.py
… aiogram/methods/send_video.py
… aiogram/methods/send_video_note.py
… aiogram/methods/send_voice.py
… aiogram/types/base.py
… aiogram/types/bot_command_scope_all_chat_administrators.py
… aiogram/types/bot_command_scope_all_group_chats.py
… aiogram/types/bot_command_scope_all_private_chats.py
… aiogram/types/bot_command_scope_chat.py
… aiogram/types/bot_command_scope_chat_administrators.py
… aiogram/types/bot_command_scope_chat_member.py
… aiogram/types/bot_command_scope_default.py
… aiogram/types/callback_query.py
… aiogram/types/chat.py
… aiogram/types/chat_invite_link.py
… aiogram/types/chat_join_request.py
… aiogram/types/chat_location.py
… aiogram/types/chat_member.py
… aiogram/types/chat_member_administrator.py
… aiogram/types/chat_member_banned.py
… aiogram/types/chat_member_left.py
… aiogram/types/chat_member_member.py