# Description

Significant performance improve about x10 more times. 
Proof of concept replace pydantic to msgspec.
Add  `aiogram.webhook.asgi.ASGIApplication` for handle webhooks

**_WARNING_**: `aiogram.filters.callback_data.CallbackData` still use pydantic need to rewrite, sorry I have not more time.

## Simple benchmark for example

**file:** start.lua
```lua
wrk.method = "POST"
wrk.body = '{"update_id":509004967,"message":{"message_id":39389,"from":{"id":79923438,"is_bot":false,"first_name":"testuser","username":"testuser","language_code":"en"},"chat":{"id":11111111,"first_name":"testuser","username":"testuser","type":"private"},"date":1683804783,"text":"/start","entities":[{"offset":0,"length":6,"type":"bot_command"}]}}'
wrk.headers['Content-Type'] = "application/json"
```

**file:** perftest.py
```python
from aiogram import Bot
from aiogram import Dispatcher
from aiogram import Router
from aiogram.enums import ParseMode
from aiogram import types
from aiogram.filters import Command

import uvloop
uvloop.install()

router = Router()

dp = Dispatcher()
bot = Bot(token='123:token', parse_mode=ParseMode.MARKDOWN_V2)
dp.include_router(router)

try:
    from aiogram.webhook.asqi import ASGIApplication
    app = ASGIApplication(bot, dp, handle_in_background=False)
except ImportError:
    print('\naiogram.webhook.asqi is not available\n used pydantic version')


@router.message(Command(commands=['start']))
async def start(message: types.Message):
    return message.reply('test passed')


if __name__ == '__main__':
    from aiohttp import web
    from aiogram.webhook.aiohttp_server import SimpleRequestHandler, setup_application
    app = web.Application()
    SimpleRequestHandler(handle_in_background=False, dispatcher=dp,
                         bot=bot).register(app, path='/test')
    setup_application(app, dp, bot=bot)
    web.run_app(app, host='127.0.0.1', port=8080)

```

### Test original version (aiohttp)
```bash
% pip uninstall aiogram
% pip install https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip
% python perftest.py

aiogram.webhook.asqi is not available.

Test aiohttp backend
======== Running on http://127.0.0.1:8080 ========
(Press CTRL+C to quit)
```

```bash
% curl --header "Content-Type: application/json" --request POST --data '{"update_id":509004967,"message":{"message_id":39389,"from":{"id":79923438,"is_bot":false,"first_name":"testuser","username":"testuser","language_code":"en"},"chat":{"id":11111111,"first_name":"testuser","username":"testuser","type":"private"},"date":1683804783,"text":"/start","entities":[{"offset":0,"length":6,"type":"bot_command"}]}}' -i http://127.0.0.1:8080/test
HTTP/1.1 200 OK
Content-Length: 867
Content-Type: multipart/form-data; boundary=webhookBoundaryde2pYoxNnBCABvVlyvMwEA
Date: Thu, 11 May 2023 13:03:03 GMT
Server: Python/3.11 aiohttp/3.8.4

--webhookBoundaryde2pYoxNnBCABvVlyvMwEA
Content-Type: text/plain; charset=utf-8
Content-Length: 11
Content-Disposition: form-data; name="method"

sendMessage
--webhookBoundaryde2pYoxNnBCABvVlyvMwEA
Content-Type: text/plain; charset=utf-8
Content-Length: 8
Content-Disposition: form-data; name="chat_id"

11111111
--webhookBoundaryde2pYoxNnBCABvVlyvMwEA
Content-Type: text/plain; charset=utf-8
Content-Length: 11
Content-Disposition: form-data; name="text"

test passed
--webhookBoundaryde2pYoxNnBCABvVlyvMwEA
Content-Type: text/plain; charset=utf-8
Content-Length: 10
Content-Disposition: form-data; name="parse_mode"

MarkdownV2
--webhookBoundaryde2pYoxNnBCABvVlyvMwEA
Content-Type: text/plain; charset=utf-8
Content-Length: 5
Content-Disposition: form-data; name="reply_to_message_id"

39389
--webhookBoundaryde2pYoxNnBCABvVlyvMwEA--
```
```bash
% wrk -t1 -c150 -d30s -s start.lua --latency http://127.0.0.1:8080/test
Running 30s test @ http://127.0.0.1:8080/test
  1 threads and 150 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   257.52ms   35.19ms 535.43ms   74.30%
    Req/Sec   584.99    206.58     1.22k    68.03%
  Latency Distribution
     50%  253.37ms
     75%  270.35ms
     90%  293.63ms
     99%  389.88ms
  17437 requests in 30.08s, 17.66MB read
Requests/sec:    579.64
Transfer/sec:    601.15KB
```

### Test improved version
```bash
% pip uninstall aiogram
% pip install https://github.com/whiteluna/aiogram/archive/refs/heads/msgspec.zip
```

#### aiohttp
```bash
% python perftest.py
Test aiohttp backend
======== Running on http://127.0.0.1:8080 ========
(Press CTRL+C to quit)
```

```bash
% curl --header "Content-Type: application/json" --request POST --data '{"update_id":509004967,"message":{"message_id":39389,"from":{"id":79923438,"is_bot":false,"first_name":"testuser","username":"testuser","language_code":"en"},"chat":{"id":11111111,"first_name":"testuser","username":"testuser","type":"private"},"date":1683804783,"text":"/start","entities":[{"offset":0,"length":6,"type":"bot_command"}]}}' -i http://127.0.0.1:8080/test
HTTP/1.1 200 OK
content-type: application/json
Content-Length: 173
Date: Thu, 11 May 2023 13:10:49 GMT
Server: Python/3.11 aiohttp/3.8.4

{"chat_id":11111111,"text":"test passed","parse_mode":"MarkdownV2","disable_web_page_preview":null,"protect_content":null,"reply_to_message_id":39389,"method":"sendMessage"

% wrk -t1 -c150 -d30s -s start.lua --latency http://127.0.0.1:8080/test
Running 30s test @ http://127.0.0.1:8080/test
  1 threads and 150 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    41.46ms    3.62ms  74.91ms   87.99%
    Req/Sec     3.62k   253.39     4.18k    74.33%
  Latency Distribution
     50%   40.55ms
     75%   41.58ms
     90%   44.84ms
     99%   57.44ms
  108338 requests in 30.08s, 32.75MB read
Requests/sec:   3601.26
Transfer/sec:      1.09MB
```

#### ASGI via uvcorn

```bash
% uvicorn --port 8080 perftest:app --no-access-log --workers 1 --loop uvloop --http httptools 
INFO:     Started server process [76993]
INFO:     Waiting for application startup.
INFO:     Application startup complete.
INFO:     Uvicorn running on http://127.0.0.1:8080 (Press CTRL+C to quit)
```

```bash
% curl --header "Content-Type: application/json" --request POST --data '{"update_id":509004967,"message":{"message_id":39389,"from":{"id":79923438,"is_bot":false,"first_name":"testuser","username":"testuser","language_code":"en"},"chat":{"id":11111111,"first_name":"testuser","username":"testuser","type":"private"},"date":1683804783,"text":"/start","entities":[{"offset":0,"length":6,"type":"bot_command"}]}}' -i http://127.0.0.1:8080/test
HTTP/1.1 200 OK
date: Thu, 11 May 2023 13:16:02 GMT
server: uvicorn
content-type: application/json
content-length: 173

{"chat_id":11111111,"text":"test passed","parse_mode":"MarkdownV2","disable_web_page_preview":null,"protect_content":null,"reply_to_message_id":39389,"method":"sendMessage"}

% wrk -t1 -c150 -d30s -s start.lua --latency http://127.0.0.1:8080/test
Running 30s test @ http://127.0.0.1:8080/test
  1 threads and 150 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    25.59ms    4.00ms  52.18ms   80.24%
    Req/Sec     5.86k   434.53     6.66k    76.67%
  Latency Distribution
     50%   25.31ms
     75%   26.07ms
     90%   27.63ms
     99%   43.49ms
  175193 requests in 30.06s, 49.96MB read
Requests/sec:   5827.23
Transfer/sec:      1.66MB
```

## Type of change

- [ ] Documentation (typos, code examples or any documentation update)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] This change requires a documentation update


# Checklist:

- [ ] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes