@JrooTJunior 2023-05-11
Also if you change any dependencies, it should be also updated in the `pyproject.toml` file.