@JrooTJunior 2023-05-11
> Please don't rush, don't release until performance issues are resolved.

all the plans are more than a year and a half behind, the expectations of the community are too long, besides, I don't have free time to continue supporting 2 versions at the same time (2.x and 3.x), and I don't agree that I shouldn't make a release in the current state.

Starting to prepare the next major release with replacement of the serializer and a number of other optimizations sooner rather than later is a more promising direction than delaying the release even more. Such a large-scale change will delay the development for 1-3 months, since everything must be carefully checked, and in the current conditions, when genocide is being committed against my country, the time for the development of the open source project is too unpredictable.