@JrooTJunior 2023-05-11
I also have high hopes for pydantic v2, which should significantly speed up its work (due to the use of Rust), but with minimal changes in the framework code, the same performance can be increased without switching to another library.

https://docs.pydantic.dev/latest/blog/pydantic-v2/
Benchmarks: https://github.com/pydantic/pydantic-core/tree/main/tests/benchmarks