@JrooTJunior 2023-07-01
So, now `pydantic` v2 is released, so i test it.

On my envieonment with v1:
```
Running 30s test @ http://127.0.0.1:8080/test
  1 threads and 150 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   154.70ms   15.68ms 208.96ms   61.70%
    Req/Sec     0.97k   180.48     1.51k    73.15%
  Latency Distribution
     50%  158.43ms
     75%  163.09ms
     90%  173.10ms
     99%  194.27ms
  29036 requests in 30.04s, 29.41MB read
Requests/sec:    966.64
Transfer/sec:      0.98MB
```

v2
```
Running 30s test @ http://127.0.0.1:8080/test
  1 threads and 150 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    57.79ms    8.96ms 118.16ms   83.32%
    Req/Sec     2.61k   282.81     3.02k    83.95%
  Latency Distribution
     50%   58.02ms
     75%   59.87ms
     90%   62.18ms
     99%   94.64ms
  77807 requests in 30.04s, 78.80MB read
Requests/sec:   2590.42
Transfer/sec:      2.62MB
```