@whiteluna 2023-07-28
But still, it remains twice as slow) Refactoring will take a lot of time, I'll close the request. Let's get back to this when there's a stable release and plenty of free time. Thanks for the work!